gitconfig: open files from git grep using fzf
Experimenting with fzf :)
This works but with one caveat : when the editor is closed, fzf is still
running.
`become` is the solution for this instead of `execute` but that hangs
for me. This comes from being a git alias because directly from the
terminal this works as expected.

See https://junegunn.github.io/fzf/reference/#command-execution